String API cleanups.
* Remove most DWORD->Pyssize_t confusion/warnings re strings in the API.
* Group the pywintypes.h string declarations together.
* Remove unused string code.

+++ win32/src/PyWinTypesmodule.cpp
@@ -376,7 +376,7 @@ static PyObject *PyWin_IsTextUnicode(PyObject *self, PyObject *args)
         return NULL;
     }
 
-    DWORD rc = IsTextUnicode((LPVOID)value, numBytes, &flags);
+    DWORD rc = IsTextUnicode((LPVOID)value, (int)numBytes, &flags);
     return Py_BuildValue("ii", rc, flags);
     // @rdesc The function returns (result, flags), both integers.
     // <nl>result is nonzero if the data in the buffer passes the specified tests.
